+++ src/BranchFlowModel.jl
@@ -1,6 +1,7 @@
 module BranchFlowModel
 
 using CommonOPF
+import CommonOPF: rij, xij  # extending to MultiPhase, need to consolidate into CommonOPF
 using JuMP
 using LinearAlgebra
 using Graphs, MetaGraphs